+++ spec/baby_squeel/relation_spec.rb
@@ -4,9 +4,7 @@ require 'shared_examples/table'
 require 'shared_examples/relation'
 
 describe BabySqueel::Relation do
-  subject(:table) {
-    BabySqueel::Relation.new(Post)
-  }
+  subject(:table) { create_relation Post }
 
   include_examples 'a table'
   include_examples 'a relation'